Remove test system definitions from ghmc_benchmark
